requirements.txt
colorama==0.4.6
dnspython==2.7.0
ipwhois==1.3.0
python-nmap==0.7.1
python-whois==0.9.5
requests==2.32.5
tabulate==0.9.0
tqdm==4.67.1
urllib3==1.26.20
python-Wappalyzer
setuptools
